Move import test to test_init
